requirements.txt
discord.py
pymongo
pillow
requests
python-dotenv
dnspython